requirements.txt
streamlit
openai
arxiv
python-dotenv
requests
tqdm
sentence-transformers
chromadb
watchdog